Added user profile
Page located at /account . Allows the user to change current username and password. Source code located at /components/MyAccount .
The main view is /components/MyAccount/MyAccount.tsx and there is subcomponents in the componentes directory.
The main components contains the logic and the needed methods, and the subcomponents contains all the validations, styling and inputs.
Currently there is three subcomponents:
- ProfileInfo.tsx: read only data from user
- UpdatePassword.tsx: password update field
- UpdateUsername.tsx: username update field